+++ src/services/token.service.ts
@@ -3,7 +3,7 @@
  */
 
 import { Contract, Signer } from 'ethers';
-import { ERC20_ABI } from '../utils/subscryptsABI';
+import { ERC20_ABI } from '../contract';
 import { ContractError, InsufficientBalanceError } from '../utils/errors';
 
 /**