+++ src/utils/index.ts
@@ -5,4 +5,3 @@
 export * from './errors';
 export * from './validators';
 export * from './formatters';
-export * from './subscryptsABI';